requirements.txt
flask
gunicorn
flask-cors
requests
numpy
python-dotenv
urllib3